--- a/src/index.html
+++ b/src/index.html
@@ -22,6 +22,7 @@
 	<script src="js/pako-inflate-2.0.4.min.js"></script>
 	<script src="data/pois.js"></script>
 	<script src="data/tilecolormap.js"></script>
+	<script src="data/mazeFilters.js"></script>
 	<script src="js/maploader.js"></script>
 	<script src="js/canvastools.js"></script>
 	<script src="js/mapmonitor.js"></script>