## Tasks
- Fixed a bug that does not work with the Show Biome feature.
- Improved tile highlighting rendering speed double.(about 580ms -> 220ms)
- Improved rendering speed independent of Show Biome or Tile Highlighting more than 40x(about 580ms -> 15ms)
- Improved Show Biome rendering speed double.(about 300ms -> 150ms)
- Refactored some codes.